+++ xtt/lib/flow/gtk/flow_draw_gtk.h
@@ -180,7 +180,7 @@ public:
   void set_cursor(FlowCtx* ctx, draw_eCursor cursor);
   void set_nav_cursor(FlowCtx* ctx, draw_eCursor cursor);
   int get_text_extent(FlowCtx* ctx, const char* text, int len,
-      flow_eDrawType gc_type, int idx, int* width, int* height, double size);
+      flow_eDrawType gc_type, int idx, double* width, double* height, double size);
   void copy_area(FlowCtx* ctx, int x, int y);
   void clear_area(FlowCtx* ctx, int ll_x, int ur_x, int ll_y, int ur_y);
   int create_input(FlowCtx* ctx, int x, int y, char* text, int len, int idx,
