+++ xtt/lib/flow/src/flow_draw.cpp
@@ -306,7 +306,7 @@ void FlowDraw::set_nav_cursor(FlowCtx* ctx, draw_eCursor cursor)
 }
 
 int FlowDraw::get_text_extent(FlowCtx* ctx, const char* text, int len,
-    flow_eDrawType gc_type, int idx, int* width, int* height, double size)
+    flow_eDrawType gc_type, int idx, double* width, double* height, double size)
 {
   return 1;
 }
